+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 
+- Named and documented a field in action 609 (`PointLight`). The field was previously named `unk_ds3_f2_24` and is now `maxViewDistance`.
 - `DataAction`s now check if there are any properties that have the wrong type or number of components when they are written to an FXR buffer, and throws descriptive errors if there are any. This should make it easier to find out if and where you put an invalid value.
 - Added the missing descriptions for the `particleLength` and `particleWidth` properties in actions 10008 and 10009.